+++ Sources/VCore/Views/SwiftUI/Fetch Delegating Completion Image/FetchDelegatingCompletionImageContent.swift
@@ -1,0 +1,11 @@
+//
+//  FetchDelegatingCompletionImageContent.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 06.07.23.
+//
+
+import Foundation
+
+// MARK: - Fetch Delegating Completion Image Content
+typealias FetchDelegatingCompletionImageContent = FetchDelegatingAsyncImageContent